Gradle Extensions for uncluttering build files
