+++ app/build.gradle.kts
@@ -1,6 +1,9 @@
 import org.jetbrains.compose.compose
 import org.jetbrains.compose.desktop.application.dsl.TargetFormat
 import com.voxfinite.logvue.Dependencies
+import com.voxfinite.logvue.utils.findPkg
+import com.voxfinite.logvue.utils.ghActionOutput
+import com.voxfinite.logvue.utils.getMainAppVersion
 
 val pluginsDir: File by rootProject.extra
 
@@ -13,7 +16,7 @@ plugins {
 val r8: Configuration by configurations.creating
 
 group = "com.voxfinite"
-version = appVersion()
+version = project.getMainAppVersion()
 val appName = "logvue"
 val appMainClass = "com.voxfinite.logvue.app.MainKt"
 
@@ -153,41 +156,4 @@ gradle.buildFinished {
     val jarPkg = buildDir.resolve("compose/jars").findPkg(".jar")
     nativePkg.ghActionOutput("app_pkg")
     jarPkg.ghActionOutput("uber_jar")
-}
-
-fun File.findPkg(format: String?) = when (format != null) {
-    true -> walk().firstOrNull { it.isFile && it.name.endsWith(format, ignoreCase = true) }
-    else -> null
-}
-
-fun File?.ghActionOutput(prefix: String) = this?.let {
-    when (System.getenv("GITHUB_ACTIONS").toBoolean()) {
-        true -> println(
-            """
-        ::set-output name=${prefix}_name::${it.name}
-        ::set-output name=${prefix}_path::${it.absolutePath}
-      """.trimIndent()
-        )
-        else -> println("$prefix: $this")
-    }
-}
-
-fun appVersion() : String {
-    val key = "APP_VERSION"
-    return if (project.hasProperty(key)) {
-        val version = project.property("APP_VERSION").toString()
-        println("Version = $version")
-        if (version.isBlank()) {
-            return "1.0.0"
-        }
-        if (version.matches(Regex("^[\\d]{1,3}.[\\d]{1,3}.[\\d]{1,4}"))) {
-            return version
-        }
-        if (version.matches(Regex("^v[\\d]{1,3}.[\\d]{1,3}.[\\d]{1,4}"))) {
-            return version.removePrefix("v")
-        }
-        "1.0.0"
-    } else {
-        "1.0.0"
-    }
 }
